third_party/WebKit/LayoutTests/platform/mac/fast/events/touch/gesture/touch-gesture-noscroll-body-propagated-expected.txt

This tests that a page cannot be scrolled vertically with touch if its body has style overflow-y:hidden and the scroll event is propogated from a scrollable child div.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first gesture
FAIL document.documentElement.scrollTop should be 0. Was 10.
PASS document.documentElement.scrollLeft is 0
PASS wheelEventsOccurred is 0
second gesture
FAIL document.documentElement.scrollTop should be 0. Was 305.
PASS document.documentElement.scrollLeft is 0
PASS wheelEventsOccurred is 0
PASS scrollEventsOccurred is 0
PASS successfullyParsed is true

TEST COMPLETE
